baselines/adaattn/pyproject.toml
[project]
name = "adaattn-baseline"
version = "0.1.0"
description = "AdaAttN: Attention-based Arbitrary Neural Style Transfer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision",
    "pillow",
    "tqdm",
]

[project.optional-dependencies]
dev = ["ruff"]

[tool.uv]
# This is a standalone project, not part of a workspace
# It has its own lockfile to handle potential dependency conflicts

